breeder-service/pyproject.toml
[tool.poetry]
name = "breeder-service"
version = "0.1.0"
description = ""
authors = ["tianqizhao-louis <louiszhao0517@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
asyncpg = "^0.29.0"
databases = {extras = ["postgresql"], version = "^0.9.0"}
sqlalchemy = "^2.0.35"
httpx = "^0.27.2"
psycopg2-binary = "^2.9.9"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
aiosqlite = "^0.20.0"
gunicorn = "^23.0.0"
gcloud-aio-pubsub = "^6.0.1"
pyjwt = "^2.10.1"
pydantic-settings = "^2.6.1"
python-logstash = "^0.4.8"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"